fix bug

+++ program/SVM/SVM/common.h
@@ -1,6 +1,6 @@
 #pragma once
 
-using input_type	= short;
+using input_type	= double;
 using mask_type		= unsigned char;
 using output_type	= short;
 using label_type	= unsigned char;
